Preserve callbacks when replacing a task with a chain (#6189)
* Preserve callbacks when replacing a task with a chain.

* Preserve callbacks when replacing a task with a chain.

* Added tests.

* Update celery/app/task.py

Co-authored-by: maybe-sybr <58414429+maybe-sybr@users.noreply.github.com>

* Mark test as flaky.

* Fix race condition in CI.

* fix: Run linked tasks in original slot for replace

This change alters the handling of linked tasks for chains which are
used as the argument to a `.replace()` call for a task which itself has
a chain of signatures to call once it completes. We ensure that the
linked callback is not only retained but also called at the appropiate
point in the newly reconstructed chain comprised of tasks from both the
replacement chain and the tail of the encapsulating chain of the task
being replaced.

We amend some tests to validate this behaviour better and ensure that
call/errbacks behave as expected if the encapsulating chain has either
set. One test is marked with an `xfail` since errbacks of encapsulating
chains are not currently called as expected due to some ambiguity in
when an errback of a replaced task should be dropped or not (#6441).

Co-authored-by: Asif Saif Uddin <auvipy@gmail.com>
Co-authored-by: maybe-sybr <58414429+maybe-sybr@users.noreply.github.com>