Add sentinel_kwargs to Rendis Sentinel docs
If the Sentinel cluster has a password, it also has to be passed down via the `sentinel_kwargs` option. If it was not supplied I got an error: `No master found for 'mymaster'`.

Google didn't do much for me trying to find this option, I found it in the source, so I think it should be documented.