test: Fix unexpected behaviour from bad mocking
This test would attempt to mock the `request_stack` of a task so as to
confirm that it could confirm that the request object pushed onto it
contained simulated delivery information as expected. However, it did
not wrap the original call target which led to an unfortunate
interaction with the worker optimisations in `app/trace.py` which would
not find the request on the stack and therefore not end up calling the
task's `run()` method.

The worker optimisations can be enabled as a side effect of other tests
like `test_regression_worker_startup_info()` in the mongo and cache
backend suites. This led to a situation where the test changed in the
diff would fail if those tests happened to run before it! Luckily, the
side effect of the worker optimizations being enabled are not what cause
the unrelated failure, the test in this diff was a just a bit unaware of
the consequences of its mocking.